Fixed compilation warning on 64-bit Windows. Fixed manipulators test to close opened file before deleting.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@50782 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/compress/api/zlib.cpp
@@ -1435,7 +1435,7 @@ void g_GZip_ScanForChunks(CNcbiIstream& is, IChunkHandler& handler)
                 }
                 // We don't need uncompressed data -- discard it
                 strm.next_out  = out_buf;
-                strm.avail_out = out_size;
+                strm.avail_out = (unsigned int)out_size;
 
                 // Decompress
                 ret = inflate(&strm, Z_SYNC_FLUSH);
